fix invalid dirty region problem
 When a ViewGroup has only one child view and the visibility of the
child view is changed from VISIBLE to INVISIBLE, then the bounds of
the ViewGroup is missed from updating dirty regions since there is no
draw operations for the ViewGroup. (Dirty region is clipped to the bounds
of a view only when it has any draw operation) Therefore, the invisible
child view remains displayed.

 This patch takes the bounds of the ViewGroup into account during
dirty region calculation in such a case.

Change-Id: Ie406d1f9df87e851fb478bff8ea0a0ef3449f2f4
Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>