Revert "DimLayer : remove unnecessary surface transaction calls"
This reverts commit 6ce3a37f6b06585c5c0c045d1df69f21f8200067.

Problem:
DimLayer is not shown during transition animation and hence
the window of an activity going background is abnormally
displayed when transition occurs.

Reason:
The patch to be reverted makes a call to adjustBounds() only
when there is an actual change of the boundary of DimLayer.
So there is the case that the requested width and height of
the surface of DimLayer are zero if there is only a change
of z-order in DimLayer. This is because the requested width
and height of a surface are maintained only during surface
transaction opened, i.e. the requested size of a surface is
cleared after closing surface transaction.

Change-Id: I26ec2d050c0b9f4502abd51c893bae7129609e5e
Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>